cdk/requirements.txt
aws-cdk-lib==2.120.0
constructs>=10.0.0,<11.0.0
boto3>=1.34.0
pydantic>=2.5.0
python-dotenv>=1.0.0
